% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_outputGenerics.r, R/GARS_outputmethods.r
\docType{methods}
\name{AllPop}
\alias{AllPop}
\alias{AllPop methods}
\alias{AllPop}
\alias{AllPop,GARS-AllPop}
\title{Accessors for the 'AllPop' slot of a GarsSelectedFeatures object.}
\usage{
AllPop(x)

\S4method{AllPop}{GarsSelectedFeatures}(x)
}
\arguments{
\item{x}{a \code{GarsSelectedFeatures} object}
}
\value{
a list containing all the populations
}
\description{
The AllPop slot contains the list of populations
}
\examples{
data(GARS_res_GA)
ex_pop <- AllPop(GARS_res_GA)
}
\author{
Mattia Chiesa, Luca Piacentini
}
