% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_outputGenerics.r, R/GARS_outputmethods.r
\docType{methods}
\name{FitScore}
\alias{FitScore}
\alias{FitScore,GarsSelectedFeatures-method}
\alias{FitScore,GARS-FitScore}
\title{Accessors for the 'FitScore' slot of a GarsSelectedFeatures object.}
\usage{
FitScore(x)

\S4method{FitScore}{GarsSelectedFeatures}(x)
}
\arguments{
\item{x}{a \code{GarsSelectedFeatures} object}
}
\value{
a vector containing the fitness scores
}
\description{
The FitScore slot contains the fitness values over the generations
}
\examples{
data(GARS_res_GA)
ex_pop <- FitScore(GARS_res_GA)
}
\author{
Mattia Chiesa, Luca Piacentini
}
