% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_createRndPopulation.R
\name{GARS.create.rnd.population}
\alias{GARS.create.rnd.population}
\title{Create a random chromosomes population}
\usage{
GARS.create.rnd.population(data, chr.len, chr.num = 1000)
}
\arguments{
\item{data}{A matrix or a data.frame of the input data.
Rows and Cols have to be, respectively, observations and features}

\item{chr.len}{The length of chromosomes. This value corresponds
to the desired length of the feature set.}

\item{chr.num}{The number of chromosomes to generate. Default is
1000}
}
\value{
A matrix representing the chromosomes population: each column
is a chromosome and each element correspond to the feature position in
'data'
}
\description{
This function creates the initial random population
of chromosomes
}
\examples{
# use example data:
data(GARS_data_norm)
GARS.create.rnd.population(GARS_data_norm, chr.len=10, chr.num=100)

}
\author{
Mattia Chiesa, Luca Piacentini
}
