% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_fitnessFunction.R
\name{GARS.FitFun}
\alias{GARS.FitFun}
\title{This function implements the Fitness Function of GARS}
\usage{
GARS.FitFun(data, classes, chr.pop)
}
\arguments{
\item{data}{A matrix or a data.frame of the input data.
Rows and Cols have to be, respectively, observations and features}

\item{classes}{A vector of type "factor" with \code{nrow(data)} elements.
Each element represents the class label for each observation.}

\item{chr.pop}{A matrix or a data.frame representing the chromosomes
population: each column is a chromosome and each element corresponds
 to the feature position in the expression data matrix}
}
\value{
A numeric vector where each element corresponds to
 the fitness score of each chromosome in 'chr.pop'
}
\description{
In GARS the Fitness Function consists in calculating the
Averaged Silhouette Index after a Multi-Dimensional Scaling
}
\examples{
# use example data:
data(GARS_data_norm)
data(GARS_classes)
data(GARS_popul)
fitness_scores <- GARS.FitFun(GARS_data_norm, GARS_classes, GARS_popul)

}
\seealso{
\code{\link{GARS.create.rnd.population}}
}
\author{
Mattia Chiesa, Luca Piacentini
}
