% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_fitevo.R
\name{GARS.PlotFitnessEvolution}
\alias{GARS.PlotFitnessEvolution}
\title{Plot the maximum fitness scores for each generation}
\usage{
GARS.PlotFitnessEvolution(fitness.scores)
}
\arguments{
\item{fitness.scores}{A numeric vector where each element corresponds
to the fitness score}
}
\value{
A plot which represent the evolution of the fitness score
across the generations
}
\description{
This function plots the maximum fitness scores for each
generation
}
\examples{
# use example data:
data(GARS_fit_list)
GARS.PlotFitnessEvolution(GARS_fit_list)

}
\seealso{
\code{\link{GARS.PlotFeaturesUsage}}
}
\author{
Mattia Chiesa, Luca Piacentini
}
