% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_packagepage.R
\docType{package}
\name{GARS}
\alias{GARS}
\alias{GARS-package}
\title{GARS package for a robust feature selection of high-dimensional data}
\description{
The main function of GARS is \code{\link{GARS_GA}}, which implements a
clustering-based Genetic Algorithm to select
Robust Subsets of features in high-dimensional datasets.
 The user can extract the results of
 \code{\link{GARS_GA}}, exploiting the assessor methods:
 \code{\link{MatrixFeatures}}, \code{\link{LastPop}},
 \code{\link{AllPop}} and \code{\link{FitScore}}.
}
\details{
See the package vignette, by typing \code{vignette("GARS")} to discover
all the \code{\link{GARS_GA}} functions.
}
\author{
Mattia Chiesa, Giada Maioli, Luca Piacentini
}
\keyword{package}
