% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_test_data.R
\docType{data}
\name{GARS_classes}
\alias{GARS_classes}
\title{RNA-seq dataset for testing GARS}
\format{A vector of type "factor" with 58 elements: 29 labelled as "N"
and 29 labelled as "T".}
\usage{
GARS_classes
}
\value{
An example data for testing \code{GARS} package
}
\description{
The class labels of the sample dataset
}
\keyword{datasets}
