% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_test_data.R
\docType{data}
\name{GARS_data_norm}
\alias{GARS_data_norm}
\title{RNA-seq dataset for testing GARS}
\format{A matrix of 157 genes (columns) and 58 samples (rows)}
\usage{
GARS_data_norm
}
\value{
An example data for testing \code{GARS} package
}
\description{
An RNA-seq normalized matrix to test several GARS functions; this dataset
was obtained using the \code{DaMirseq} package to normalize the raw count
matrix present in \code{MLSeq} package.
}
\keyword{datasets}
