% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_test_data.R
\docType{data}
\name{GARS_fit_list}
\alias{GARS_fit_list}
\title{RNA-seq dataset for testing GARS}
\format{A numeric vector with 100 fitness scores}
\usage{
GARS_fit_list
}
\value{
An example data for testing \code{GARS} package
}
\description{
A numeric vector with the maximum fitness score for each iteration
}
\keyword{datasets}
