% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_test_data.R
\docType{data}
\name{GARS_Fitness_score}
\alias{GARS_Fitness_score}
\title{RNA-seq dataset for testing GARS}
\format{A numeric vector with 50 fitness scores}
\usage{
GARS_Fitness_score
}
\value{
An example data for testing \code{GARS} package
}
\description{
A numeric vector with the fitness scores for each chromosome in a
single generation
}
\keyword{datasets}
