% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_test_data.R
\docType{data}
\name{GARS_pop_list}
\alias{GARS_pop_list}
\title{RNA-seq dataset for testing GARS}
\format{A list with 100 consecutive chromosomes populations}
\usage{
GARS_pop_list
}
\value{
An example data for testing \code{GARS} package
}
\description{
A list containing 100 of consecutive chromosomes populations
}
\keyword{datasets}
