% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_test_data.R
\docType{data}
\name{GARS_popul}
\alias{GARS_popul}
\title{RNA-seq dataset for testing GARS}
\format{A matrix of 20 rows (features) and 50 columns (chromosomes)}
\usage{
GARS_popul
}
\value{
An example data for testing \code{GARS} package
}
\description{
A matrix to test several GARS functions, representing a
chromosome population
}
\keyword{datasets}
