% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_test_data.R
\docType{data}
\name{GARS_res_GA}
\alias{GARS_res_GA}
\title{A GarsSelectedFeatures object for testing GARS}
\format{A GarsSelectedFeatures}
\usage{
GARS_res_GA
}
\value{
An example data for testing \code{GARS} package
}
\description{
An object representing the output of GARS_GA
}
\keyword{datasets}
