% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARS_outputGenerics.r, R/GARS_outputmethods.r
\docType{methods}
\name{LastPop}
\alias{LastPop}
\alias{LastPop methods}
\alias{LastPop}
\alias{LastPop,GARS-LastPop}
\title{Accessors for the 'LastPop' slot of a GarsSelectedFeatures object.}
\usage{
LastPop(x)

\S4method{LastPop}{GarsSelectedFeatures}(x)
}
\arguments{
\item{x}{a \code{GarsSelectedFeatures} object}
}
\value{
a matrix containing the last population
}
\description{
The LastPop slot contains the last chromosome population
}
\examples{
data(GARS_res_GA)
ex_pop <- LastPop(GARS_res_GA)
}
\author{
Mattia Chiesa, Luca Piacentini
}
